move cyrillic implementation to generic.rs
