+++ src/alphabets/cyrillic.rs
@@ -1,101 +1,17 @@
-use std::cmp::Reverse;
-
 use super::RawOutcome;
-use crate::alphabets::generic::{get_all_chars_in_langs, get_lang_chars, is_relevant_for_langs};
+use crate::alphabets::generic;
 use crate::core::{FilterList, LowercaseText};
-use crate::{Lang, Script};
-
-const BUL: &str = "абвгдежзийклмнопрстуфхцчшщъьюя";
-const RUS: &str = "абвгдежзийклмнопрстуфхцчшщъыьэюяё";
-const UKR: &str = "абвгдежзийклмнопрстуфхцчшщьюяєіїґ";
-const BEL: &str = "абвгдежзйклмнопрстуфхцчшыьэюяёіў";
-const SRP: &str = "абвгдежзиклмнопрстуфхцчшђјљњћџ";
-const MKD: &str = "абвгдежзиклмнопрстуфхцчшѓѕјљњќџ";
-
-const ALL: &str = "абвгдежзийклмнопрстуфхцчшщъыьэюяёєіїґўђјљњћџѓѕќ";
+use crate::Script;
 
 pub fn alphabet_calculate_scores(text: &LowercaseText, filter_list: &FilterList) -> RawOutcome {
-    let mut raw_scores: Vec<(Lang, i32)> = Script::Cyrillic
-        .langs()
-        .iter()
-        .filter(|&&l| filter_list.is_allowed(l))
-        .map(|&l| (l, 0i32))
-        .collect();
-
     let all_langs = Script::Cyrillic.langs();
-    let all_chars_in_langs = get_all_chars_in_langs(all_langs);
-
-    // let max_raw_score = text.chars().filter(|&ch| is_relevant(ch)).count();
-    let max_raw_score = text
-        .chars()
-        .filter(|&ch| is_relevant_for_langs(&ch, &all_chars_in_langs))
-        .count();
-
-    for (lang, score) in &mut raw_scores {
-        let alphabet = get_lang_chars(*lang);
-
-        for ch in text.chars() {
-            // if !is_relevant(ch) {
-            if !is_relevant_for_langs(&ch, &all_chars_in_langs) {
-                continue;
-            } else if alphabet.contains(&ch) {
-                *score += 1;
-            } else {
-                *score -= 1;
-            }
-        }
-    }
-
-    raw_scores.sort_unstable_by_key(|(_, score)| Reverse(*score));
-
-    let raw_scores: Vec<(Lang, usize)> = raw_scores
-        .into_iter()
-        .map(|(l, s)| {
-            let score = if s < 0 { 0usize } else { s as usize };
-            (l, score)
-        })
-        .collect();
-
-    let mut normalized_scores = vec![];
-
-    for &(lang, raw_score) in &raw_scores {
-        // avoid devision by zero
-        let normalized_score = if raw_score == 0 {
-            0.0
-        } else {
-            raw_score as f64 / max_raw_score as f64
-        };
-        normalized_scores.push((lang, normalized_score));
-    }
-
-    RawOutcome {
-        count: max_raw_score,
-        raw_scores,
-        scores: normalized_scores,
-    }
+    generic::alphabet_calculate_scores_generic(text, filter_list, all_langs)
 }
-
-// fn is_relevant(ch: char) -> bool {
-//     ALL.chars().any(|c| c == ch)
-// }
-//
-// fn get_lang_chars(lang: Lang) -> Vec<char> {
-//     let alphabet = match lang {
-//         Lang::Bul => BUL,
-//         Lang::Rus => RUS,
-//         Lang::Ukr => UKR,
-//         Lang::Bel => BEL,
-//         Lang::Srp => SRP,
-//         Lang::Mkd => MKD,
-//
-//         _ => panic!("No alphabet for {}", lang),
-//     };
-//     alphabet.chars().collect()
-// }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::Lang;
 
     const CYRILLIC_LANGS: [Lang; 6] = [
         Lang::Rus,
